Implement UpdateNews as Symfony command
